Normal and Mesh fixes
Fixed the normals and extruded some planes in both the tutorial and in Jimmy's level. Some parts may require feedback from the programming / implementation team to get just right.

Jimmy's level is very detailed with no separation. Currently unable to find a good way to separate all of the objects effectively and in a way that is conducive to texturing.